Move os-specific functions.
