+++ bu/BuPyramid.cpp
@@ -13,6 +13,7 @@
 #include "OctantInfo.hpp"
 #include "../untwine/Common.hpp"
 #include "../untwine/ProgressWriter.hpp"
+#include "dirlist.hpp"
 
 namespace untwine
 {
@@ -53,7 +54,7 @@ void BuPyramid::getInputFiles()
         return std::make_pair(true, VoxelKey(x, y, z, level));
     };
 
-    std::vector<std::string> files = directoryList(m_b.opts.tempDir);
+    std::vector<std::string> files = os::directoryList(m_b.opts.tempDir);
 
     VoxelKey root;
     for (std::string file : files)
