+++ bu/PointAccessor.hpp
@@ -14,6 +14,7 @@
 
 #include "../untwine/Common.hpp"
 #include "../untwine/Point.hpp"
+#include "mapfile.hpp"
 
 #include "FileInfo.hpp"
 
@@ -31,13 +32,13 @@ public:
     ~PointAccessor()
     {
         for (FileInfo *fi : m_fileInfos)
-            unmapFile(fi->context());
+            os::unmapFile(fi->context());
     }
 
     void read(FileInfo& fi)
     {
         std::string filename = m_b.opts.tempDir + "/" + fi.filename();
-        auto ctx = mapFile(filename, true, 0, fi.numPoints() * m_b.pointSize);
+        auto ctx = os::mapFile(filename, true, 0, fi.numPoints() * m_b.pointSize);
         if (ctx.m_addr == nullptr)
             throw FatalError(filename + ": " + ctx.m_error);
         fi.setContext(ctx);
