+++ epf/Epf.cpp
@@ -33,6 +33,7 @@
 #include <pdal/util/FileUtils.hpp>
 #include <pdal/util/ProgramArgs.hpp>
 
+#include "dirlist.hpp"
 
 namespace untwine
 {
@@ -251,6 +252,10 @@ void Epf::fillMetadata(const pdal::PointLayout& layout)
         m_b.offset[1] = calcOffset(m_b.trueBounds.miny, m_b.trueBounds.maxy, m_b.scale[1]);
         m_b.offset[2] = calcOffset(m_b.trueBounds.minz, m_b.trueBounds.maxz, m_b.scale[2]);
     }
+
+    std::cerr << "Scale = " << m_b.scale[0] << "/" << m_b.scale[1] << "/" << m_b.scale[2] << "!\n";
+    std::cerr << "Offset = " << m_b.offset[0] << "/" << m_b.offset[1] << "/" << m_b.offset[2] << "!\n";
+    exit(0);
 }
 
 void Epf::createFileInfos(const StringList& input, std::vector<FileInfo>& fileInfos)
@@ -265,7 +270,7 @@ void Epf::createFileInfos(const StringList& input, std::vector<FileInfo>& fileIn
     {
         if (FileUtils::isDirectory(filename))
         {
-            StringList dirfiles = directoryList(filename);
+            StringList dirfiles = os::directoryList(filename);
             filenames.insert(filenames.end(), dirfiles.begin(), dirfiles.end());
         }
         else
@@ -436,9 +441,11 @@ std::vector<FileInfo> Epf::processLas(pdal::LasReader& r, FileInfo fi)
     fi.bounds = h.getBounds();
     fi.numPoints = h.pointCount();
 
+
+std::cerr << "Scale x/y/z = " << h.scaleX() << "/" << h.scaleY() << "/" << h.scaleZ() << "!\n";
     m_b.scale[0] = (std::max)(m_b.scale[0], h.scaleX());
-    m_b.scale[1] = (std::max)(m_b.scale[0], h.scaleY());
-    m_b.scale[2] = (std::max)(m_b.scale[0], h.scaleZ());
+    m_b.scale[1] = (std::max)(m_b.scale[1], h.scaleY());
+    m_b.scale[2] = (std::max)(m_b.scale[2], h.scaleZ());
 
     fi.offsets[0] = h.offsetX();
     fi.offsets[1] = h.offsetY();
