+++ epf/Reprocessor.cpp
@@ -18,6 +18,7 @@
 
 #include "Reprocessor.hpp"
 #include "../untwine/Common.hpp"
+#include "mapfile.hpp"
 
 namespace untwine
 {
@@ -51,7 +52,7 @@ Reprocessor::Reprocessor(const VoxelKey& k, int numPoints, int pointSize,
 
 void Reprocessor::run()
 {
-    auto ctx = mapFile(m_filename, true, 0, m_fileSize);
+    auto ctx = os::mapFile(m_filename, true, 0, m_fileSize);
     if (ctx.addr() == nullptr)
     {
         std::cerr << "FATAL: " + m_filename + ": " + ctx.what();
@@ -69,7 +70,7 @@ void Reprocessor::run()
         cell->advance();
         pos += m_pointSize;
     }
-    unmapFile(ctx);
+    os::unmapFile(ctx);
     pdal::FileUtils::deleteFile(m_filename);
 }
