+++ untwine/Untwine.cpp
@@ -22,6 +22,8 @@
 
 #include "../epf/Epf.hpp"
 #include "../bu/BuPyramid.hpp"
+#include "dirlist.hpp"
+#include "stringconv.hpp"
 
 namespace untwine
 {
@@ -114,7 +116,7 @@ void cleanup(const std::string& dir, bool rmdir)
     std::regex re("[0-9]+-[0-9]+-[0-9]+-[0-9]+.bin");
     std::smatch sm;
 
-    const std::vector<std::string>& files = directoryList(dir);
+    const std::vector<std::string>& files = os::directoryList(dir);
     for (const std::string& f : files)
         if (std::regex_match(f, sm, re))
             pdal::FileUtils::deleteFile(dir + "/" + f);
@@ -136,7 +138,7 @@ int main(int argc, char *argv[])
     argv++;
     argc--;
     while (argc--)
-        arglist.push_back(untwine::fromNative(*argv++));
+        arglist.push_back(untwine::os::fromNative(*argv++));
 
     using namespace untwine;
