+++ untwine/mingw/stringconv.hpp
@@ -1,0 +1,1 @@
+#include "../generic/stringconv.hpp"
